requirements.txt
click==8.1.7
gitdb==4.0.11
GitPython==3.1.40
joblib==1.3.2
regex==2023.10.3
smmap==5.0.1
tqdm==4.66.1
prometheus-client==0.18.0
